refactor: move shipping form tests
